requirements.txt
fastapi==0.108.0
uvicorn==0.27.0
google-generativeai==0.5.0
pydantic==2.5.3
python-dotenv==1.0.0
requests==2.31.0
python-multipart==0.0.6
jinja2==3.1.2
cachetools==5.3.2
typing-extensions==4.9.0
starlette==0.31.1
redis==5.0.1
prometheus-client==0.19.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.25
email-validator==2.1.0
psycopg2-binary==2.9.9
PyJWT==2.8.0 